src: Implement an Address Space Controller(ASC)
Many of the PPU actions are governed by writes and reads to specific
addresses. Currently, we can't have code triggered by a write or read to
a specific address.

Implement an Address Space Controller(ASC) to handle writes and reads to
an address space, and with that, redirect the access to specific
addresses to Ppu variables.

Fix #8