Change the project name to "Rodomo"
After very long time thinking with my girlfriend about what would be the
name of this emulator, we found a suitable name.

This name comes from the Portuguese phrase "Rodou, mo!", the literal
translation is "It worked, love!". I always say this to my girlfriend
when something works on the emulator.

Fix #7